to change ID type of ProductRepository.java

+++ src/main/java/kz/trankwilizator/tafl/dao/product/ProductRepository.java
@@ -8,6 +8,6 @@ import org.springframework.security.access.annotation.Secured;
 
 @Secured("MODERATOR")
 @RepositoryRestResource(path="products")
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-  
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
 }
